Replaced the expiry list with a GQueue data structure.
Expiry list in the GInetFlowTable uses a GList. It will take a long
time if we add a new element at the end of the list which might
contain 500,000 elements.  Using a GQueue will trim down the time
to add to the bottom of the queue. Then make sure to allocate and
free the queues.

Also added a unit test for expire queues.